Vastly improve the cacheability of page publication queries
Previously page publication queries were made to the microsecond.  Given that an admin user will enter a date time in the format of DD/MM/YYYY HH:MM, it makes no sense to query at such a granular level.  By removing the seconds and microseconds we allow for page publication queries to be valid for 1 minute at a time.